feat : sothey extract, transform 마무리 및 코드 정리
