--- transform/transform.py
@@ -1,11 +1,0 @@
-
-# 1. 결측치 제거
-# 2. 각각 년도별로 group by
-# 3. 정규분포를 그린후 하위 30%, 상위 30% 가격 제거
-# 4. 각각의 년도별로 남은 데이터들 중에서 평균 구하기
-
-
-def getAveragePricePerYear():
-    pass
-
-getAveragePricePerYear()
